+++ opensource/kernel-cpp/kram/sosarray.h
@@ -205,7 +205,7 @@ inline T& Sos_array<T>::operator[]( int4 index ) const
 template< class T >
 inline Sos_simple_array<T>::Sos_simple_array()
 :
-    _increment ( max( 1u, 100 / sizeof( T ))),
+    _increment ( max( (size_t)1, 100 / sizeof( T ))),
     _array     ( 0 ),
     _size      ( 0 )
 {
